appengine/requirements.txt
Flask==1.0.2
Flask-BasicAuth
google-cloud-firestore
